do not delete tombstonables

+++ assembl/views/api2/__init__.py
@@ -54,7 +54,7 @@ from assembl.auth import (
 from assembl.auth.util import get_roles, get_permissions
 from assembl.semantic.virtuoso_mapping import get_virtuoso
 from assembl.models import (
-    AbstractIdeaVote, User, DiscussionBoundBase, Discussion)
+    AbstractIdeaVote, User, DiscussionBoundBase, Discussion, HistoryMixin)
 from assembl.lib.decl_enums import DeclEnumType
 
 FIXTURE_DIR = os.path.join(
@@ -313,7 +313,11 @@ def instance_del(request):
     instance = ctx._instance
     if not instance.user_can(user_id, CrudPermissions.DELETE, permissions):
         return HTTPUnauthorized()
-    instance.db.delete(instance)
+    if isinstance(instance, HistoryMixin):
+        instance.is_tombstone = True
+    else:
+        instance.db.delete(instance)
+    # maybe instance.tombstone() ?
     return {}
